updated db for postgres

+++ migrations/20190112101746_goals.js
@@ -2,9 +2,9 @@ exports.up = function(knex, Promise) {
   return knex.schema.createTable("goals", table => {
     table.increments("id");
     table.string("area");
-    table.integer("healthy");
-    table.integer("low");
-    table.integer("high");
+    table.float("healthy");
+    table.float("low");
+    table.float("high");
   });
 };
